add back minimized downstairs/src/main.rs
